cs, xaml 파일 통합 및 Top으로 그래프 시각화 기능변경
