+++ Hamtory_WPF/HistoryReader.xaml
@@ -1,32 +1,64 @@
-﻿<Window x:Class="Hamtory_WPF.MainWindow2"
+﻿<Window x:Class="Hamtory_WPF.HistoryReader"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        Title="History Data Reader" Height="600" Width="800">
+        xmlns:lc="clr-namespace:LiveCharts.Wpf;assembly=LiveCharts.Wpf"
+        Title="History Data Reader" Height="800" Width="1000">
     <Grid>
-        <StackPanel Orientation="Horizontal" HorizontalAlignment="Center" VerticalAlignment="Top" Margin="0,10,0,0">
-            <Label Content="Start Date:" VerticalAlignment="Center" Margin="5"/>
-            <DatePicker x:Name="datePickerStart" Width="120" Height="25" VerticalAlignment="Center" Margin="5"/>
-            <TextBox x:Name="txtStartTime" Width="50" Height="25" VerticalAlignment="Center" Margin="5" Text="00:00"/>
-            <Label Content="End Date:" VerticalAlignment="Center" Margin="15,0,5,0"/>
-            <DatePicker x:Name="datePickerEnd" Width="120" Height="25" VerticalAlignment="Center" Margin="5"/>
-            <TextBox x:Name="txtEndTime" Width="50" Height="25" VerticalAlignment="Center" Margin="5" Text="23:59"/>
-            <Button x:Name="btnFilter" Content="Filter" Width="100" Height="30" Click="BtnFilter_Click" RenderTransformOrigin="1.213,0.658"/>
-        </StackPanel>
-        <DataGrid x:Name="dataGrid" HorizontalAlignment="Left" VerticalAlignment="Top" Margin="10,60,10,0" AutoGenerateColumns="False" Width="760" Height="400">
-            <DataGrid.Columns>
-                <DataGridTextColumn Header="Date" Binding="{Binding STD_DT}" Width="*"/>
-                <DataGridTextColumn Header="Num" Binding="{Binding NUM}" Width="*"/>
-                <DataGridTextColumn Header="Melt Temp" Binding="{Binding MELT_TEMP}" Width="*"/>
-                <DataGridTextColumn Header="Motor Speed" Binding="{Binding MOTORSPEED}" Width="*"/>
-                <DataGridTextColumn Header="Melt Weight" Binding="{Binding MELT_WEIGHT}" Width="*"/>
-                <DataGridTextColumn Header="INSP" Binding="{Binding INSP}" Width="*"/>
-                <DataGridTextColumn Header="TAG" Binding="{Binding TAG}" Width="*"/>
-            </DataGrid.Columns>
-        </DataGrid>
-        <StackPanel Orientation="Horizontal" HorizontalAlignment="Center" VerticalAlignment="Bottom" Margin="0,10,0,10">
-            <Button Name="btnShowStatistics" Content="Statistics" Width="150" Height="30" Margin="10,0,0,0" Click="BtnShowStatistics_Click"/>
-            <Button Name="btnShowNGDistribution" Content="NG분포도" Width="150" Height="30" Margin="10,0,0,0" Click="BtnShowNGDistribution_Click"/>
-            <Button Name="btnSaveData" Content="Save Data" Width="150" Height="30" Margin="10,0,0,0" Click="BtnSaveData_Click"/>
-        </StackPanel>
+        <TabControl>
+            <!-- Data Reader 탭 -->
+            <TabItem Header="Data Reader">
+                <Grid>
+                    <StackPanel Orientation="Horizontal" HorizontalAlignment="Center" VerticalAlignment="Top" Margin="0,10,0,0">
+                        <Label Content="Start Date:" VerticalAlignment="Center" Margin="5"/>
+                        <DatePicker x:Name="datePickerStart" Width="120" Height="25" VerticalAlignment="Center" Margin="5"/>
+                        <TextBox x:Name="txtStartTime" Width="50" Height="25" VerticalAlignment="Center" Margin="5" Text="00:00"/>
+                        <Label Content="End Date:" VerticalAlignment="Center" Margin="15,0,5,0"/>
+                        <DatePicker x:Name="datePickerEnd" Width="120" Height="25" VerticalAlignment="Center" Margin="5"/>
+                        <TextBox x:Name="txtEndTime" Width="50" Height="25" VerticalAlignment="Center" Margin="5" Text="23:59"/>
+                        <Button x:Name="btnFilter" Content="Filter" Width="100" Height="30" Margin="5,0,0,0" Click="BtnFilter_Click"/>
+                        <Button x:Name="btnSaveData" Content="Save Data" Width="100" Height="30" Margin="5,0,0,0" Click="BtnSaveData_Click"/>
+                    </StackPanel>
+                    <DataGrid x:Name="dataGrid" HorizontalAlignment="Left" VerticalAlignment="Top" Margin="10,60,10,0" AutoGenerateColumns="False" Width="960" Height="600">
+                        <DataGrid.Columns>
+                            <DataGridTextColumn Header="Date" Binding="{Binding STD_DT}" Width="*"/>
+                            <DataGridTextColumn Header="Num" Binding="{Binding NUM}" Width="*"/>
+                            <DataGridTextColumn Header="Melt Temp" Binding="{Binding MELT_TEMP}" Width="*"/>
+                            <DataGridTextColumn Header="Motor Speed" Binding="{Binding MOTORSPEED}" Width="*"/>
+                            <DataGridTextColumn Header="Melt Weight" Binding="{Binding MELT_WEIGHT}" Width="*"/>
+                            <DataGridTextColumn Header="INSP" Binding="{Binding INSP}" Width="*"/>
+                            <DataGridTextColumn Header="TAG" Binding="{Binding TAG}" Width="*"/>
+                        </DataGrid.Columns>
+                    </DataGrid>
+                </Grid>
+            </TabItem>
+            <!-- Statistics 탭 -->
+            <TabItem Header="Statistics">
+                <Grid>
+                    <Grid.RowDefinitions>
+                        <RowDefinition Height="*"/>
+                        <RowDefinition Height="*"/>
+                        <RowDefinition Height="*"/>
+                        <RowDefinition Height="*"/>
+                    </Grid.RowDefinitions>
+                    <lc:CartesianChart Name="meltTempChart" Grid.Row="0"/>
+                    <lc:CartesianChart Name="motorSpeedChart" Grid.Row="1"/>
+                    <lc:CartesianChart Name="meltWeightChart" Grid.Row="2"/>
+                    <lc:CartesianChart Name="inspChart" Grid.Row="3"/>
+                </Grid>
+            </TabItem>
+            <!-- NG Distribution 탭 -->
+            <TabItem Header="NG 분포도">
+                <Grid>
+                    <StackPanel HorizontalAlignment="Center" VerticalAlignment="Center" Margin="0,10,0,0">
+                        <TextBlock Text="NG 분포도" FontSize="20" FontWeight="Bold" HorizontalAlignment="Center"/>
+                        <lc:PieChart Name="ngPieChart" Width="500" Height="300" HorizontalAlignment="Center" VerticalAlignment="Center" Margin="0,10,0,0"/>
+                        <StackPanel Orientation="Horizontal" HorizontalAlignment="Center" Margin="0,10,0,0">
+                            <TextBlock Name="txtOKCount" FontSize="16" Margin="10"/>
+                            <TextBlock Name="txtNGCount" FontSize="16" Margin="10"/>
+                        </StackPanel>
+                    </StackPanel>
+                </Grid>
+            </TabItem>
+        </TabControl>
     </Grid>
 </Window>
