--- Hamtory_WPF/NGDistributionWindow.xaml
@@ -1,16 +1,0 @@
-﻿<Window x:Class="Hamtory_WPF.NGDistributionWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:lc="clr-namespace:LiveCharts.Wpf;assembly=LiveCharts.Wpf"
-        Title="NG Distribution" Height="400" Width="600">
-	<Grid>
-		<StackPanel HorizontalAlignment="Center" VerticalAlignment="Top" Margin="0,10,0,0">
-			<TextBlock Text="NG 분포도" FontSize="20" FontWeight="Bold" HorizontalAlignment="Center"/>
-			<lc:PieChart Name="ngPieChart" Width="500" Height="300" Margin="0,10,0,0"/>
-			<StackPanel Orientation="Horizontal" HorizontalAlignment="Center" Margin="0,10,0,0">
-				<TextBlock Name="txtOKCount" FontSize="16" Margin="10"/>
-				<TextBlock Name="txtNGCount" FontSize="16" Margin="10"/>
-			</StackPanel>
-		</StackPanel>
-	</Grid>
-</Window>
