--- Hamtory_WPF/StatisticsWindow.xaml
@@ -1,18 +1,0 @@
-﻿<Window x:Class="Hamtory_WPF.StatisticsWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:lc="clr-namespace:LiveCharts.Wpf;assembly=LiveCharts.Wpf"
-        Title="Statistics Graph" Height="800" Width="1000">
-	<Grid>
-		<Grid.RowDefinitions>
-			<RowDefinition Height="*"/>
-			<RowDefinition Height="*"/>
-			<RowDefinition Height="*"/>
-			<RowDefinition Height="*"/>
-		</Grid.RowDefinitions>
-		<lc:CartesianChart Name="meltTempChart" Grid.Row="0"/>
-		<lc:CartesianChart Name="motorSpeedChart" Grid.Row="1"/>
-		<lc:CartesianChart Name="meltWeightChart" Grid.Row="2"/>
-		<lc:CartesianChart Name="inspChart" Grid.Row="3"/>
-	</Grid>
-</Window>
